requirements.txt
alembic==1.4.2
certifi==2020.6.20
chardet==3.0.4
#cplex==12.10.0.0
#docloud==1.0.375
#docplex==2.11.176
geographiclib==1.50
geopy==1.20.0
haversine==2.1.2
idna==2.7
Mako==1.1.3
MarkupSafe==1.1.1
nltk==3.3
numpy==1.19.1
opencage==1.2
pandas==0.25.3
plotly==4.5.4
psycopg2==2.8.5
PyPDF2==1.26.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
requests==2.20.0
retrying==1.3.3
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.24.3
